Split module test_base into module base and module test_checker
